scraper/requirements.txt
playwright==1.41.0
pytest==7.4.4
python-dateutil==2.8.2
beautifulsoup4==4.12.2
lxml==4.9.3
azure-cosmos==4.5.0
azure-core==1.35.0
python-dotenv==1.0.0
requests==2.32.5
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0